+++ tests/test_clean.rs
@@ -1,5 +1,5 @@
 use clap::Parser;
-use code_it_later_rs::fs_operation::run_format_command_to_file;
+//use code_it_later_rs::fs_operation::run_format_command_to_file;
 use code_it_later_rs::{args::*, *};
 use std::fs::{self, copy, remove_file};
 use std::io::{prelude::*, BufReader, Result};
@@ -37,25 +37,25 @@ fn test_clean_the_crumbs() -> Result<()> {
 
 //#[test]
 //:= fmt command change, this test case need to fix in future
-fn test_fmt_after_clean() -> Result<()> {
-    copy(
-        "./tests/testcases/format_test.go.bkp",
-        "./tests/testcases/format_test.go",
-    )?;
+// fn test_fmt_after_clean() -> Result<()> {
+//     copy(
+//         "./tests/testcases/format_test.go.bkp",
+//         "./tests/testcases/format_test.go",
+//     )?;
 
-    let args = Args::parse_from(vec!["codeitlater", "./tests/testcases/format_test.go"]);
-    let conf = config::Config::from(&args);
+//     let args = Args::parse_from(vec!["codeitlater", "./tests/testcases/format_test.go"]);
+//     let conf = config::Config::from(&args);
 
-    let mut bread = fs_operation::handle_files(conf);
-    let file_path = fs_operation::clean_the_crumbs(bread.next().unwrap())?;
+//     let mut bread = fs_operation::handle_files(conf);
+//     let file_path = fs_operation::clean_the_crumbs(bread.next().unwrap())?;
 
-    assert!(run_format_command_to_file("go fmt", vec![file_path]).is_ok());
+//     assert!(run_format_command_to_file("go fmt", vec![file_path]).is_ok());
 
-    assert!(same_file(
-        "tests/testcases/format_test.go.expect",
-        "tests/testcases/format_test.go",
-    )
-    .unwrap());
+//     assert!(same_file(
+//         "tests/testcases/format_test.go.expect",
+//         "tests/testcases/format_test.go",
+//     )
+//     .unwrap());
 
-    remove_file("./tests/testcases/format_test.go")
-}
+//     remove_file("./tests/testcases/format_test.go")
+// }