+++ tests/test_keywords.rs
@@ -13,7 +13,6 @@ fn test_keywords() {
         "target",
         "-k",
         "TODO",
-        "--",
         "./tests/testcases/keywords.lisp",
     ]);
 
@@ -29,6 +28,62 @@ fn test_keywords() {
                 Some("TODO".to_string()),
                 "this is TODO".to_string()
             ),]
+        )]
+    );
+}
+
+#[test]
+fn test_ignore_keyword_file() {
+    let args = Args::parse_from(vec![
+        "codeitlater",
+        "-x",
+        "target",
+        "-k",
+        "TODO",
+        "./tests/testcases/test.rs",
+    ]);
+
+    let conf = config::Config::from(&args);
+
+    assert_eq!(
+        fs_operation::handle_files(conf).collect::<Vec<_>>(),
+        vec![Bread::new(
+            "./tests/testcases/test.rs".to_string(),
+            vec![Crumb::new(
+                6,
+                0,
+                Some("TODO".to_string()),
+                "this is the ignore line".to_string()
+            )
+            .add_ignore_flag()]
+        )]
+    );
+
+    let args = Args::parse_from(vec![
+        "codeitlater",
+        "-x",
+        "target",
+        "-k",
+        "MARK",
+        "./tests/testcases/keywords.lisp",
+    ]);
+
+    let conf = config::Config::from(&args);
+
+    assert_eq!(
+        fs_operation::handle_files(conf).collect::<Vec<_>>(),
+        vec![Bread::new(
+            "./tests/testcases/keywords.lisp".to_string(),
+            vec![
+                Crumb::new(3, 0, Some("MARK".to_string()), "this is MARK".to_string()),
+                Crumb::new(
+                    4,
+                    0,
+                    Some("MARK".to_string()),
+                    "this is ignored MARK".to_string()
+                )
+                .add_ignore_flag()
+            ]
         )]
     );
 }