+++ tests/testcases/keywords.lisp
@@ -1,5 +1,6 @@
 ;;;;;;;;:= TODO: this is TODO
 
 ;;:= MARK: this is MARK
+;;:= !MARK: this is ignored MARK
 
 ;;:= SOMETHING-ELSE: something else